django_app/ml_app/serializers.py
from rest_framework import serializers
from .models import LabeledData

class LabeledDataSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = LabeledData
        fields = ['text', 'label']
