django_app/requirements.txt
Django==3.2.14
djangorestframework==3.13.1
requests==2.18.4
mysqlclient
whitenoise
PyMySQL[rsa]